mon-app/backend/requirement.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.7

# Traitement de texte et ML
scikit-learn==1.3.0
numpy==1.24.3
nltk==3.8.1
pandas==2.0.3

# Traitement de documents
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.9.0
docx2txt==0.8
pdf2image==1.16.3

# OCR (optionnel)
pytesseract==0.3.10
Pillow==10.0.0

# Base de données et utilitaires
sqlite3==3.42.0  # Inclus avec Python
python-magic==0.4.27
requests==2.31.0

# Tests
pytest==7.4.0
pytest-flask==1.2.0

# Développement
python-dotenv==1.0.0